Safari-609.1.11.1/LayoutTests/fast/inspector-support/uncaught-dom8-exception.html

<html>
<head>
<script>

function onload()
{
    if (window.testRunner)
        testRunner.dumpAsText();
    document.body.removeAttributeNode(document.createAttribute("doesNotExist"));
}

</script>
</head>

<body onload="onload()">
<p>
This tests that we are getting the correct message for DOM Exception 8: NOT_FOUND_ERR.
</p>

</body>
</html>
